allow specifying display name for pinned resources
